Processing prototype/starting point (fixed again)
I made a really simple version of what our program should do.  It just
takes 2 already provided images and calculates the difference. I think
it looks promising so we can build off of it.  There's a lot more stuff
that needs to be done, as documented in the Processing source code